+++ common/src/main/java/net/mehvahdjukaar/polytone/content/particle/custom/render/ModelParticleRenderGroup.java
@@ -1,11 +1,11 @@
 package net.mehvahdjukaar.polytone.content.particle.custom.render;
 
 import net.mehvahdjukaar.polytone.content.particle.custom.CustomParticleInstance;
-import net.mehvahdjukaar.polytone.content.particle.custom.CustomParticleType;
 import net.minecraft.CrashReport;
 import net.minecraft.CrashReportCategory;
 import net.minecraft.ReportedException;
 import net.minecraft.client.Camera;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.particle.Particle;
 import net.minecraft.client.particle.ParticleEngine;
 import net.minecraft.client.particle.ParticleGroup;
@@ -25,7 +25,7 @@ public class ModelParticleRenderGroup extends ParticleGroup<CustomParticleInstan
     @Override
     public ParticleGroupRenderState extractRenderState(Frustum frustum, Camera camera, float f) {
         for (CustomParticleInstance particle : this.particles) {
-            if (particleInFrustum(frustum, particle)) {
+            if (particleInFrustum(frustum, particle, camera)) {
                 try {
                     particle.extractModel(this.particleTypeRenderState, camera, f);
                 } catch (Throwable var9) {
@@ -41,10 +41,17 @@ public class ModelParticleRenderGroup extends ParticleGroup<CustomParticleInstan
     }
 
 
-    public static boolean particleInFrustum(Frustum frustum, Particle particle) {
+    public static boolean particleInFrustum(Frustum frustum, Particle particle, Camera camera) {
         AABB bb = particle.getBoundingBox();
-        return frustum.intersection.testAab((float) (bb.minX - frustum.camX), (float) (bb.minY - frustum.camY), (float) (bb.minZ - frustum.camZ),
+        boolean inFrustum = frustum.intersection.testAab((float) (bb.minX - frustum.camX), (float) (bb.minY - frustum.camY), (float) (bb.minZ - frustum.camZ),
                 (float) (bb.maxX - frustum.camX), (float) (bb.maxY - frustum.camY), (float) (bb.maxZ - frustum.camZ));
+        if (camera.entity() == Minecraft.getInstance().player) {
+            //vista?
+            if (particle instanceof CustomParticleInstance cpt) {
+                cpt.notifyInFrustum(inFrustum);
+            }
+        }
+        return inFrustum;
     }
 
 }