+++ common/src/main/java/net/mehvahdjukaar/polytone/mixins/QuadParticleGroupMixin.java
@@ -2,6 +2,7 @@ package net.mehvahdjukaar.polytone.mixins;
 
 import com.llamalad7.mixinextras.sugar.Local;
 import net.mehvahdjukaar.polytone.content.particle.custom.render.ModelParticleRenderGroup;
+import net.minecraft.client.Camera;
 import net.minecraft.client.particle.QuadParticleGroup;
 import net.minecraft.client.particle.SingleQuadParticle;
 import net.minecraft.client.renderer.culling.Frustum;
@@ -15,7 +16,8 @@ public class QuadParticleGroupMixin {
     @Redirect(method = "extractRenderState",
             require = 0,
             at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/culling/Frustum;pointInFrustum(DDD)Z"))
-    public boolean polytone$accurateFrustumIntersect(Frustum instance, double d, double e, double f, @Local SingleQuadParticle particle) {
-        return ModelParticleRenderGroup.particleInFrustum(instance, particle);
+    public boolean polytone$accurateFrustumIntersect(Frustum instance, double d, double e, double f,
+                                                     @Local SingleQuadParticle particle, @Local(argsOnly = true) Camera camera) {
+        return ModelParticleRenderGroup.particleInFrustum(instance, particle, camera);
     }
 }