+++ changelog.md
@@ -1,1 +1,1 @@
-- added an aggressive, on by default, particle rate limiter. might worsen the look of some packs (or make them way faster if they are not optimized). turn off for best looks.
+- added aggressive, on by default "kill_when_not_in_view" parameter to custom particles that will kill a particle when not in the render frustum